requirements.txt
Django          
djangorestframework
markdown       
django-filter
channels[daphne]
channels-redis

psycopg2-binary
httpx
elasticsearch
celery
redis
django-cors-headers
django-extensions
djangorestframework-simplejwt
requests
phonenumbers
ipython
drf-yasg

